requirements.txt
python-dotenv
fastapi
uvicorn[standard]
sqlalchemy
sqlmodel
alembic
itsdangerous
PyJWT

passlib
bcrypt==4.0.1
python-multipart
zipstream-new

psycopg2
pandas

google-auth-oauthlib
google-auth-httplib2
sib_api_v3_sdk
redis
colorlog
asyncpg
pydantic-settings
pydantic[email]
httpx
resend
boto3
simpy
